Refactor navigation component.
