+++ src/components/Header/Nav/NavItem/NavItem.jsx
@@ -1,0 +1,9 @@
+import { NavLink } from "react-router-dom";
+
+export function NavItem({ link }) {
+  return (
+    <li>
+        <NavLink to={link.toLowerCase().trim()}>{link}</NavLink>
+    </li>
+  )
+}
